+++ src/SQLQueryStress/SQLQueryStress.csproj
@@ -11,6 +11,7 @@
     <PublisherName>Adam Machanic</PublisherName>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <RunAnalyzersDuringBuild>true</RunAnalyzersDuringBuild>
+    <StartupObject>SQLQueryStress.Program</StartupObject>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <RunCodeAnalysis>true</RunCodeAnalysis>
@@ -36,9 +37,9 @@
     <Compile Update="DatabaseSelect.Designer.cs">
       <DependentUpon>DatabaseSelect.cs</DependentUpon>
     </Compile>
-    <Compile Update="Form1.cs" />
-    <Compile Update="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Update="FormMain.cs" />
+    <Compile Update="FormMain.Designer.cs">
+      <DependentUpon>FormMain.cs</DependentUpon>
     </Compile>
     <Compile Update="DataViewer.cs" />
     <Compile Update="DataViewer.Designer.cs">
@@ -63,9 +64,9 @@
       <SubType>Designer</SubType>
       <DependentUpon>DatabaseSelect.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Update="Form1.resx">
+    <EmbeddedResource Update="FormMain.resx">
       <SubType>Designer</SubType>
-      <DependentUpon>Form1.cs</DependentUpon>
+      <DependentUpon>FormMain.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Update="DataViewer.resx">
       <SubType>Designer</SubType>
